[US_IX] Set validation thresholds and revise validations for US_IX (Recidiviz/recidiviz-data#19039)
## Description of the change

We just turned off dev mode for US_IX validations, so this PR
temporarily bumps some validation thresholds for failing validations due
to one of the following reasons:
- These validations are currently failing at a very low rate. Therefore,
I bumped the threshold for these and left tickets open for them in
**Icebox** for future investigation after we've finished resolving
higher priority validation failures:
- `most_recent_face_to_face_contact_date_by_person_by_state_comparison`
  - `admission_pfi_pop_pfi_mismatch`
  - `incarceration_commitments_subset_of_admissions`
  - `incarceration_release_reason_no_release_date`
- These sessions validations are likely connected to known sentence
dates issues that we're still actively working through. Therefore, I
bumped the thresholds for these, and left tickets open for them in
**Prioritized Backlog** to revisit once sentence dates are finalized on
the ingest side and we've resolved the other sentence date related
failing validations.
  - `session_new_admissions_with_no_sentence_date_imposed`
  - `session_liberty_releases_with_no_sentence_completion_date`

3/3/2023: For the rest of the failing validations, I've just bumped the
thresholds for now to stop them from hard failing, but am
mid-investigation or still planning on investigating in the near future.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 024dee1393237c600990c548f720461efe586caa