add debug/default parameters to PipelineParameters class and template jsons (Recidiviz/recidiviz-data#19275)
## Description of the change

* add debug and default parameters to PipelineParameters class and
template_metadata.json files for each type of pipeline.
* allow the pipeline code to handle default values (instead of setting
calculation_month_count to -1 as a default in PipelineParameters, just
do not pass in that parameter to the flex template if it is not set)
* reverted the PipelineParameters class to manually return keys for
`_template_parameter_keys`. In the airflow environment, reading a
filepath from root to get the keys from the template_metadata.json files
was not working as expected, so instead of trying to make it work I
decided to return keys manually instead. i think this also makes more
sense because any update (addition or deletion of a parameter) to the
template_metadata.json parameter fields would require an update to the
PipelineParameters class as well (adding or deleting a field from the
class).

in addition,
## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#19118

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 8aae13ee74816be22acda98b4176793239afb22b