[Ingest] Hydrate update_timestamp params at query building time (Recidiviz/recidiviz-data#19257)
## Description of the change

There's no real reason we need to build ingest view queries that still
have `@param` parameters in them when we can just manually inject a
`DATETIME()` clause wherever we need a datetime bound.

In this PR I do away with `@update_datetime` query params (and
associated param override functionality) and just update the ingest view
building code to inject the dates directly.

This refactor is important to do now because right now we categorize raw
data queries as either PARAMETERIZED or LATEST.

However, LATEST will not work when we have ingest view queries that
query ALL the rows in a given raw data table, not just the collapsed
latest rows. I want to clarify the concepts so it's clear we're talking
about taking a shortcut and querying from a view that is already
deployed (can only be used when there is no date bound) or when we're
expanding out the raw data query to use the proper logic we need.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Pre-work for Recidiviz/recidiviz-data#18282.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d6c0f2561d2fd726a6146ddcf4b7b22344c3233a