[US_MO] Update supervision period view to correctly remove duplicate periods (Recidiviz/recidiviz-data#19282)
## Description of the change

The overlapping_supervision_periods validation reveals a number of
supervision periods in MO with the same start and end date. This is
occurring because of the order of CTEs in the supervision view
query–currently, duplicate periods are getting removed in
`field_assignments_with_unique_date_spans`, and open periods are
subsequently closed in `augmented_field_assignments`. This means that if
closing a period results in a duplicate, it will not be caught. This
change switches the order of the CTEs such that open periods are closed,
and _then_ duplicate periods are removed.

This change makes it so that FIELD_ASSIGNMENT_SEQ_NUM is assigned at the
end of the field assignment CTEs, so that the sequence number is
actually sequential and won't have gaps caused by removed periods.
FIELD_ASSIGNMENT_SEQ_NUM is a part of the supervision periods' external
ID, so note that certain external IDs will be altered by this change.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#18897

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 47f64ed8a8ee6d9cf821d585b20074ecb2348842