Updating everything to a BranchPythonOperator (Recidiviz/recidiviz-data#19355)
## Description of the change

Instead of using ShortCircuitOperators, we should update everything to
use BranchPythonOperators. This should hopefully allow for a better
understanding of mechanics and also I tested this out in the Airflow
experiment environment the behavior is as expected.

You can see the entire graph at
https://cd55e3ffc08c42488e232db689cfd60a-dot-us-central1.composer.googleusercontent.com/dags/recidiviz-staging_sftp_dag/graph?root=

## Type of change

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

SFTP fixes

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1ab63562045aa1afdc303e297edd13c11228f4bd