Adds state assignments to officer_assignments (Recidiviz/recidiviz-data#19420)
## Description of the change

This PR passes state assignments through to the officer level to have
them hydrate in `experiments.officer_assignments`. We could instead use
a view that accomplishes the same task, but the materialization time is
~5 seconds, which could be annoying for purposes like Looker, and the
additional storage is in the MB range (i.e. trivial cost).

Following this PR landing in staging we'll need to update the Looker
view `officer_assignments` to pull _only_ the materialized table and not
union state-level assignments again.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes None but helpful for Recidiviz/recidiviz-data#19362

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 1990615c7daceccfcd3d28f874efafffb21fceff