Client record exclusions (Recidiviz/recidiviz-data#19430)
## Description of the change

This PR does the following:
- Adds a CTE in `client_record_ctes` to exclude people who aren't
showing up in raw data. This only works for states where we get full
historical transfers each time, which is TN and MI for now
- Adds the same CTE to a TN-specific client record file,
`supervision_clients_template` which will ultimately be deprecated as
part of the CR refactor work
- Removes the logic that was doing this for TN in opportunity record
queries

The following query compares sandbox output to a sandbox off main. As
expected, TN expiration and supervision level downgrade look the same
(since we were already doing this exclusion in opportunity record
queries) and no other changes since MI is not yet in client record.

```
WITH sandbox AS (
  select state_code, all_eligible_opportunities2, count(*) as sandbox_count
  from `recidiviz-staging.dsharm_20230314_workflows_views.client_record_materialized`,
  unnest(all_eligible_opportunities) AS all_eligible_opportunities2
  group by 1,2
  order by 1,2
), main AS (
    select state_code, all_eligible_opportunities2, count(*) as main_count
  from `recidiviz-staging.dsharm_20230314_main_workflows_views.client_record_materialized`,
  unnest(all_eligible_opportunities) AS all_eligible_opportunities2
  group by 1,2
  order by 1,2
)
SELECT *
FROM sandbox
inner join main
  using(state_code, all_eligible_opportunities2)

```

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 09a6f81bb6f25cef7d7ca7aead498fee23ac8db1