Fix `populate_is_deleted_column_in_raw_data_tables` to add `is_deleted` to tmp schema (Recidiviz/recidiviz-data#19521)
## Description of the change

When running the script in staging, I realized that we actually _did_
need to sometimes update the tmp table schema to include `is_deleted`
for tables that had been populated in BQ before a deploy adding
`is_deleted` as a nullable column. It was crashing with a `404` when
copying from an `is_deleted` column in a table that didn't have it yet
in original tables that were in this in between state.

Fixing this by adding `is_deleted` as a column in temporary tables for
any tables that did _not_ already have it in the schema.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Recidiviz/recidiviz-data#18954

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 31455d9559741c0633d0d2d1858fea72daf27ceb